requirements.txt
awsebcli==3.10.6
blessed==1.14.2
botocore==1.7.16
cement==2.8.2
colorama==0.3.7
Django==1.9.12
django-npm==1.0.0
django-webpack-loader==0.6.0
djangorestframework==3.6.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
npm==0.1.1
optional-django==0.1.0
pathspec==0.5.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.44.0
